refactor(field_required): add field required to controllers

+++ app/controllers/user.js
@@ -1,4 +1,3 @@
-import assert from 'assert';
 import { User, Blog, Comment } from '../models';
 
 export default {
@@ -21,7 +20,10 @@ export default {
     const where = { id };
     const user = await User.findOne({ where });
 
-    assert(user, '没有此用户');
+    if (!user) {
+      ctx.failToJson(404, '没有此用户');
+      return;
+    }
 
     await Blog.destroy({ where: { userId: id } });
     await Comment.destroy({ where: { userId: id } });
